Timestamp is int64

+++ hydra/buffer_test.go
@@ -10,7 +10,7 @@ import (
 
 func newDummyRecordSet(n int) *fluent.FluentRecordSet {
 	records := make([]fluent.FluentRecordType, n)
-	ts := uint64(time.Now().Unix())
+	ts := time.Now().Unix()
 	for i := 0; i < n; i++ {
 		data := make(map[string]interface{})
 		data["message"] = []byte(fmt.Sprintf("message%d", i))
